A timecard-like Google Chart rendering, based off of the GitHub punch cards, like the one at:   http://github.com/cyberfox/jbidwatcher/graphs/punch_card
This requires the cyberfox-gchart gem (0.5.3) from GitHub, based off
the original (Rubyforge) gchart gem with some modifications to support
this kind of scatter graph.  The standard gchart gem is not suitable,
unfortunately, because of fundamental design decisions.

Specifically scaling the points based on the highest value (x or y) in
the entire graph makes it hard to have a graph 'canvas' that's larger
than your actual point use.  Also, it's not clear to me that scaling
all data sets based on the max value in all of them (x, y, or size,
for example) is correct.  Rather than change the basic design, I
just use 'text' instead of 'extended' data types, and fill the data
values in as they're given, not scaled at all.